+++ pubtools/_impl/pluggy.py
@@ -1,13 +1,40 @@
 from __future__ import absolute_import
 
 import sys
+import logging
 from contextlib import contextmanager
 
+import pkg_resources
 import pluggy
+
+LOG = logging.getLogger("pubtools")
 
 pm = pluggy.PluginManager("pubtools")
 hookspec = pluggy.HookspecMarker("pubtools")
 hookimpl = pluggy.HookimplMarker("pubtools")
+
+
+def resolve_hooks():
+    # A private helper to ensure all code defining hookspecs/hookimpls has been imported
+    # before continuing.
+
+    # 1. Eagerly load any pubtools console_scripts entry points.
+    #
+    # This will pick up hookspecs from task libraries such as pubtools-quay.
+    #
+    for ep in pkg_resources.iter_entry_points("console_scripts"):
+        if ep.module_name.startswith("pubtools"):
+            ep.resolve()
+            LOG.debug("Resolved %s", ep)
+
+    # 2. Eagerly load any pubtools.hooks entry points.
+    #
+    # This is a group we provide so that any hook-only modules, which might otherwise
+    # not be imported by anyone, can request themselves to be imported.
+    #
+    for ep in pkg_resources.iter_entry_points("pubtools.hooks"):
+        ep.resolve()
+        LOG.debug("Resolved %s", ep)
 
 
 @hookspec
@@ -45,9 +72,12 @@ def task_context():
 
     The context manager will ensure that:
 
+    * hookspecs/hookimpls are resolved across all installed libraries.
     * :func:`task_start` is invoked when the block is entered.
     * :func:`task_stop` is invoked when the block is exited.
     """
+    resolve_hooks()
+
     pm.hook.task_start()
 
     failed = True